GetEbusInfo (no bus station)

+++ socketSceneLoader/EBus.h
@@ -28,7 +28,7 @@ public :
 	string time;
 	Status status;
 	VType vtype;
-	pair<int, int> location;
+	pair<float, float> location;
 	pair<int, int> direction;// ¡ý(0,1) 1  ¡ú(1,0) 2  ¡ü(0,-1) 3  ¡û(-1,0) 4
 	float speed;
 	bool isShowEnergy;
